backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.5
psycopg2-binary>=2.9
requests>=2.31
pandas>=2.1
openpyxl>=3.1
djangorestframework-simplejwt>=2.3
python-dotenv>=1.0.0
